requirements.txt
discord.py
pypresence
rich